grpc_service/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "neighborhood-library-grpc"
version = "0.1.0"
description = "Internal gRPC service for Neighborhood Library"
requires-python = ">=3.11"
dependencies = [
  "grpcio>=1.78",
  "grpcio-health-checking>=1.78",
  "grpcio-tools>=1.78",
  "pymongo>=4.6",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
